refactor: 알림 생성 UI 파일 구조 변경 및 텍스트 필드 추가
알림 관련 UI 파일들을 view 폴더로 이동하고 알림 생성 화면에
텍스트 필드를 추가한다.

- 파일 경로 변경: `lib/presentation/notification/ui` 에서
  `lib/presentation/notification/ui/view` 로 이동
- 알림 생성 화면에 제목 텍스트와 텍스트 필드 추가
- 관련 위젯 파일 추가: `notification_text`,
  `notification_text_field`
- SingleChildScrollView 추가: 화면 오버플로우 방지
