+++ lib/presentation/notification/ui/view/check_notification_screen.dart
